Minioサービスのヘルスチェックを追加し、ユーザー画像編集機能を実装
