+++ next/src/app/[locale]/(common-layout)/user/[handle]/edit/_components/user-edit-action.ts
@@ -4,7 +4,7 @@ import { getCurrentUser, unstable_update } from "@/auth";
 import { parseFormData } from "@/lib/parse-form-data";
 import { redirect } from "next/navigation";
 import { z } from "zod";
-import { updateUser } from "./_db/mutations.server";
+import { updateUser } from "../_db/mutations.server";
 import reservedHandles from "./reserved-handles.json";
 const RESERVED_HANDLES = [...new Set([...reservedHandles])];
 const schema = z.object({
