+++ next/src/app/[locale]/(common-layout)/user/[handle]/edit/_components/user-image-edit-action.test.ts
@@ -2,7 +2,7 @@ import { uploadImage } from "@/app/[locale]/_lib/upload";
 import { getCurrentUser, unstable_update } from "@/auth";
 import { mockUsers } from "@/tests/mock";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { updateUserImage } from "./_db/mutations.server";
+import { updateUserImage } from "../_db/mutations.server";
 import { userImageEditAction } from "./user-image-edit-action";
 
 vi.mock("@/auth", () => ({
