+++ next/src/app/[locale]/(common-layout)/user/[handle]/edit/_components/user-image-edit-action.ts
@@ -3,7 +3,7 @@ import { uploadImage } from "@/app/[locale]/_lib/upload";
 import type { ActionResponse } from "@/app/types";
 import { getCurrentUser, unstable_update } from "@/auth";
 import { redirect } from "next/navigation";
-import { updateUserImage } from "./_db/mutations.server";
+import { updateUserImage } from "../_db/mutations.server";
 
 export type UserImageEditState = ActionResponse<
 	{
